requirements.txt
psutil~=5.9.5
comtypes~=1.2.0
pygrabber~=0.1
pywin32~=306
numpy~=1.25.1
Pillow~=10.0.0
opencv-python~=4.8.0.74
winotify~=1.1.0
mediapipe~=0.10.1
PySide6~=6.5.1.1
PyQt6~=6.5.1
joblib~=1.3.1
scikit-learn~=1.3.0
pystray~=0.19.4